[UI]: moved UI classes files to 'UI/' folder.
